% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bam_managment.R
\name{add_bams}
\alias{add_bams}
\title{Add BAMs to Sample Sheet}
\usage{
add_bams(zent_obj, alignment_dir)
}
\arguments{
\item{zent_obj}{ZentTools object.}

\item{alignment_dir}{Directory of aligned reads.}
}
\description{
Add BAMs to Sample Sheet
}
