% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_peaks.R
\name{annotate_peaks}
\alias{annotate_peaks}
\title{Annotate Peaks}
\usage{
annotate_peaks(
  zent_obj,
  outdir = getwd(),
  genome_annotation,
  promoter_downstream = 1000,
  promoter_upstream = 1000,
  feature_type = "gene"
)
}
\arguments{
\item{zent_obj}{ZentTools object.}

\item{outdir}{Directory to output the annotated peaks.}

\item{genome_annotation}{Genome GTF annotation file.}

\item{promoter_downstream}{Bases downstream of TSS to define promoter.}

\item{promoter_upstream}{Bases upstream of TSS to define promoter.}

\item{feature_type}{Either 'gene' or 'transcript'.}
}
\description{
Annotate Peaks
}
