% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_count.R
\name{count_features}
\alias{count_features}
\title{Count Features}
\usage{
count_features(
  zent_obj,
  outdir = getwd(),
  count_feature = "exon",
  aggregate_feature = "gene_id",
  multiple_overlap = FALSE,
  minimum_overlap = 10,
  largest_overlap = TRUE,
  read_extension_3prime = 0,
  multi_mapping = FALSE,
  multi_mapping_frac = FALSE,
  strand_specific = 0
)
}
\arguments{
\item{zent_obj}{Zent object.}

\item{outdir}{Output directory for counts.}

\item{count_feature}{Reads will be counted against this
genomic feature type, such as the default 'exon' (GTF.featureType).}

\item{aggregate_feature}{Features such as 'exons' will have
their counts aggregated based on this feature type, such as
the default 'gene_id' (GTF.attrType).}

\item{multiple_overlap}{Whether to allow reads to overlap with more
than one feature.}

\item{minimum_overlap}{The minimum number of bases that the
read needs to overlap to be counted.}

\item{largest_overlap}{Whether the read will be assigned
to the feature it has the most overlapping bases with.}

\item{read_extension_3prime}{The number of bases to extend the read
downstream from the 3' end.}

\item{multi_mapping}{Whether to count multi-mapping reads.}

\item{multi_mapping_frac}{Whether multi-mapping reads are
counted as fractions.}

\item{strand_specific}{Either 0 (unstranded), 1 (stranded),
or 2 (reverse stranded).}
}
\description{
Count Features
}
