% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_control.R
\name{fastqc}
\alias{fastqc}
\title{Quality Control of Fastq Files}
\usage{
fastqc(zent_obj, outdir = getwd())
}
\arguments{
\item{zent_obj}{Zent object.}

\item{outdir}{Output directory for reports (-o).}
}
\description{
Quality Control of Fastq Files
}
