% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_seqs.R
\name{get_seqs}
\alias{get_seqs}
\title{Get Genomic Sequences}
\usage{
get_seqs(zent_obj, outdir = getwd(), fixed_width = NA)
}
\arguments{
\item{zent_obj}{ZentTools object.}

\item{outdir}{Output directory for fasta file.}

\item{fixed_width}{Make the peaks a fixed width from the center of the peak.}
}
\description{
Get Genomic Sequences
}
