% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigwigs.R
\name{make_bigwigs}
\alias{make_bigwigs}
\title{Generate Bigwigs}
\usage{
make_bigwigs(
  zent_obj,
  outdir = getwd(),
  bin_size = 10,
  normalize_using = NA,
  genome_size = NA,
  skip_non_covered = TRUE,
  min_fragment = NA,
  max_fragment = NA,
  extend_reads = 200
)
}
\arguments{
\item{zent_obj}{ZentTools object.}

\item{outdir}{Output directory.}

\item{bin_size}{Bin size for coverage summary.}

\item{normalize_using}{Either 'CPM' or 'RPGC'.}

\item{genome_size}{Effective genome size.}

\item{skip_non_covered}{Should regions without coverage be skipped.}

\item{min_fragment}{Minimum fragment length.}

\item{max_fragment}{Maximum fragment length.}

\item{extend_reads}{Distance to extend single-end reads.}
}
\description{
Generate Bigwigs
}
