% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setting_handler.R
\name{pull_setting}
\alias{pull_setting}
\title{Pull Settings}
\usage{
pull_setting(zent_obj, setting)
}
\arguments{
\item{zent_obj}{ZentTools object.}

\item{setting}{Setting to pull.}
}
\description{
Pull Settings
}
