% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve.R
\name{retrieve_reads}
\alias{retrieve_reads}
\title{Retrieve Reads from SRA}
\usage{
retrieve_reads(zent_obj, outdir = getwd())
}
\arguments{
\item{zent_obj}{Zent object.}

\item{outdir}{Directory to donwload files to.}
}
\description{
Retrieve Reads from SRA
}
