% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setting_handler.R
\name{set_settings}
\alias{set_settings}
\title{Set Settings.}
\usage{
set_settings(
  zent_obj,
  analysis_type = NA,
  paired = NA,
  ncores = NA,
  genome_dir = NA,
  genome_annotation = NA,
  genome_assembly = NA,
  alignment_dir = NA,
  peak_dir = NA
)
}
\arguments{
\item{zent_obj}{ZentTools object.}

\item{analysis_type}{Type of experiment.
Either 'RNA-seq', 'ChIP-seq', or 'ChEC-seq'.}

\item{paired}{Paired-end status of the reads.
Either TRUE or FALSE.}

\item{ncores}{The number of cores/threads to use.}

\item{genome_dir}{The directory of the genome index.}

\item{genome_annotation}{The directory and file name of the
the genome index GTF file.}

\item{genome_assembly}{The directory and file name of the
the gnome assembly FASTA file.}

\item{alignment_dir}{The directory containing the aligned reads.}

\item{peak_dir}{The directory containing the called peaks.}
}
\description{
Set Settings.
}
