% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment.R
\name{star_genome}
\alias{star_genome}
\title{STAR Genome Generation}
\usage{
star_genome(
  zent_obj,
  outdir = getwd(),
  genome_assembly,
  genome_annotation,
  sa_bases = NA
)
}
\arguments{
\item{zent_obj}{Zent object.}

\item{outdir}{Output directory for STAR genome index (--genomeDir).}

\item{genome_assembly}{Path to genome fasta file (--genomeFastaFiles).}

\item{genome_annotation}{Path to genome gtf file (--sjdbGTFfile).}

\item{sa_bases}{Number of suffix array bases (--genomeSAindexNbases)}
}
\description{
STAR Genome Generation
}
