% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/star_alignment.R
\name{star_index}
\alias{star_index}
\title{STAR Genome Generation}
\usage{
star_index(
  zent_obj,
  outdir = getwd(),
  genome_assembly,
  genome_annotation,
  sa_bases = NA,
  ram_limit = NA
)
}
\arguments{
\item{zent_obj}{Zent object.}

\item{outdir}{Output directory for STAR genome index.}

\item{genome_assembly}{Path to genome fasta file.}

\item{genome_annotation}{Path to genome gtf file.}

\item{sa_bases}{Number of suffix array base.}

\item{ram_limit}{Upper memory limit for genome generate.}
}
\description{
STAR Genome Generation
}
