% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zentobj.R
\docType{class}
\name{zent_obj-class}
\alias{zent_obj-class}
\title{Zent Class}
\description{
Zent Class
}
\section{Slots}{

\describe{
\item{\code{sample_sheet}}{Sample sheet.}

\item{\code{settings}}{Various run settings.}
}}

