% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zentobj.R
\name{zent_tools}
\alias{zent_tools}
\title{Zent Object Constructor Function}
\usage{
zent_tools(
  analysis_type,
  sample_sheet,
  sep = "\\t",
  paired = FALSE,
  ncores = 1
)
}
\arguments{
\item{analysis_type}{Either 'RNA-seq' or 'ChIP-seq'.
Only 'RNA-seq' is currently available.}

\item{sample_sheet}{Either a data.frame or delimited file.}

\item{sep}{If the sample sheet is a file, this specifies the delimiter.}

\item{paired}{Whether the run is paired (TRUE) or unpaired (FALSE)}

\item{ncores}{The number of cores/threads to use.}
}
\value{
A zent object.
}
\description{
Zent Object Constructor Function
}
