% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigwigs.R
\name{make_bigwigs}
\alias{make_bigwigs}
\title{Generate Bigwigs}
\usage{
make_bigwigs(
  zent_obj,
  outdir = getwd(),
  bin_size = 10,
  normalize_using = NA,
  genome_size = NA,
  skip_non_covered = TRUE,
  min_fragment = NA,
  max_fragment = NA,
  extend_reads = 200,
  scale_factors = NA,
  split_strands = FALSE,
  library_type = NA
)
}
\arguments{
\item{zent_obj}{ZentTools object.}

\item{outdir}{Output directory.}

\item{bin_size}{Bin size for coverage summary.}

\item{normalize_using}{Either 'CPM' or 'RPGC'.}

\item{genome_size}{Effective genome size.}

\item{skip_non_covered}{Should regions without coverage be skipped.}

\item{min_fragment}{Minimum fragment length.}

\item{max_fragment}{Maximum fragment length.}

\item{extend_reads}{Distance to extend single-end reads.
Set to NA to not extend reads.}

\item{scale_factors}{Takes a named vector, with the name being the sample name,
and the value being the scale factor.
If set will override 'normalzie_using' option.}

\item{split_strands}{For RNA-seq, whether to split the strands into
positive and minus strand files.}

\item{library_type}{If split_strands is TRUE, specify library chemistry
as either 'dUTP' or 'ligation'.}
}
\description{
Generate Bigwigs
}
