% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/star_alignment.R
\name{star_align}
\alias{star_align}
\title{STAR Alignment}
\usage{
star_align(zent_obj, outdir = getwd())
}
\arguments{
\item{zent_obj}{Zent object.}

\item{outdir}{Output directory for aligned reads.
For use with '--outFileNamePrefix'.}
}
\description{
STAR Alignment
}
